--- src/types/page-tracker.ts
@@ -1,4 +1,0 @@
-import type { Pageview } from "../api/pageview";
-import type { Screenview } from "../api/screenview";
-
-export type PageTrackerParams = Pageview | Screenview;
