--- src/types/router.ts
@@ -1,4 +1,0 @@
-export type {
-  RouteLocationNormalizedGeneric as Route,
-  Router,
-} from "vue-router";
